+++ NEWS.md
@@ -3,6 +3,7 @@
 - `SaveDatatable` functions now run more efficiently and can handle larger 
   Datatables
 - Added `methods` to NAMESPACE to allow Rscript to run checks on the package
+- Added type checks for `SaveDatatable`
 
 ## BUG FIXES